feat: 데이터베이스에 저장 구현
