+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
 spring.application.name=spring-gift
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:mem:product
+spring.datasource.username=sa
+spring.datasource.password=
