+++ src/main/resources/schema.sql
@@ -1,0 +1,8 @@
+create table products
+(
+    id       bigint auto_increment,
+    name     varchar(100),
+    price    int,
+    imageUrl varchar(255),
+    primary key (id)
+);
